Endpoint: Change state when endpoint Regenerate fails.
When the endpoint regeneration fails, the status of that endpoint will
be `StateWaitingToRegenerate` and the new regeneration will be lock
becasue the readyToRegenerate always returns false:

```
readyToRegenerate := e.SetStateLocked(StateWaitingToRegenerate, "Triggering regeneration due to new identity")
```

The main reason, the SetStateLocked does not allow update status from
StateWaitingToRegenerate to StateWaitingToRegenerate, so the endpoint
will be in lock in the same status.

With this change, if the endpoint fails the status will change to
regenerating, the correct status should be "not-ready" or something
similar. The log messages are a bit weird right now.

```
2018-06-22T07:41:02Z   Failure   ready                   Error regenerating endpoint: Unable to rename current endpoint directory: rename /var/run/cilium/state/18260 /var/run/cilium/state/18260_stale: no such file or directory
```

Related to  #4541 and #4598

Signed-off-by: Eloy Coto <eloy.coto@gmail.com>