KVStore: refactor and add metrics.
This only adds a new functionality based on metrics, but the refactor the
kvstore is a bit large.

With this change `kvstore.go` now does not export any method, all the methods
are now under `ExtendedKVStore` that wraps a namespace, that enables to know
who is using the kvstore to send metrics.

Now all `BackendOperations` methods that interact with the kvstore need to have
a namespace argument, to send the metrics. The main reason to add the namespaces is
because that methods are public, so can be used by any other package. (As
example, kvstore/store/)

New metrics are the following:

```
kvstore_operations_total                            kind="read" scope="allocator" action="GetCapabilities"   1.000000
kvstore_operations_total                            action="ListPrefix" kind="read" scope="allocator"        3.000000
kvstore_operations_total                            kind="update" scope="ipcache" action="Update"            12.000000
kvstore_operations_total                            kind="update" scope="shared-store" action="Update"       28.000000
```

The reason of the labels:

- Scope, that it's the namespace, to know what package uses the kvstore.
- Kind: There are a few kind (delete, read, set) that defines what action was
  made in the kvstore integrations.
- action: This is a low level information. is not too big (12 actions) and it's
  mainly to debug in cases that the kvstore is heavily used.

The idea for the metric is:

- Have a dashboard with the kind of operation
- Have a debug dashboard that display how many interactions per scope.

This commit is related to #5382

Signed-off-by: Eloy Coto <eloy.coto@gmail.com>